Added: proper casting,gate config in SR1 file

+++ flamedisx/lz/lz_WS2024.py
@@ -325,16 +325,16 @@ class LZ24ERSource(LZWS2024Source, fd.nest.nestERSource):
             Constants are direct over-rides of eqn 6 in Arxiv: 2211.10726 
             Energy: energy in keV
         """
-        m1=12.4886
-        m2=85.0
-        m3=0.6050
-        m4= 2.14687
-        m5=25.721
-        m6=0. 
-        m7=59.651
-        m8=3.6869
-        m9=0.2872
-        m10=0.1121 
+        m1=tf.cast(12.4886,tf.float32)
+        m2=tf.cast(85.0,tf.float32)
+        m3=tf.cast(0.6050,tf.float32)
+        m4= tf.cast(2.14687,tf.float32)
+        m5=tf.cast(25.721,tf.float32)
+        m6=tf.cast(0. ,tf.float32)
+        m7=tf.cast(59.651,tf.float32)
+        m8=tf.cast(3.6869,tf.float32)
+        m9=tf.cast(0.2872,tf.float32)
+        m10=tf.cast(0.1121 ,tf.float32)
 
         Nq = energy  / self.Wq_keV  #equation is in keV   
 
@@ -436,18 +436,18 @@ class LZ24NRSource(LZWS2024Source, fd.nest.nestNRSource):
             See section C. in Arxiv: 2211.10726 
             Energy: energy in keV
         """
-        nr_nuis_alpha = 10.19
-        nr_nuis_beta = 1.11
-        nr_nuis_gamma = 0.0498
-        nr_nuis_delta = -0.0533
-        nr_nuis_epsilon = 12.46
-        nr_nuis_zeta =  0.2942
-        nr_nuis_eta = 1.899
-        nr_nuis_theta = 0.3197
-        nr_nuis_l = 2.066
-        nr_nuis_p = 0.509
-        nr_new_nuis_a = 0.996
-        nr_new_nuis_b =  0.999
+        nr_nuis_alpha = tf.cast(10.19,tf.float32)
+        nr_nuis_beta = tf.cast(1.11,tf.float32)
+        nr_nuis_gamma = tf.cast(0.0498,tf.float32)
+        nr_nuis_delta = tf.cast(-0.0533,tf.float32)
+        nr_nuis_epsilon = tf.cast(12.46,tf.float32)
+        nr_nuis_zeta =  tf.cast(0.2942,tf.float32)
+        nr_nuis_eta = tf.cast(1.899,tf.float32)
+        nr_nuis_theta = tf.cast(0.3197,tf.float32)
+        nr_nuis_l = tf.cast(2.066,tf.float32)
+        nr_nuis_p = tf.cast(0.509,tf.float32)
+        nr_new_nuis_a = tf.cast(0.996,tf.float32)
+        nr_new_nuis_b =  tf.cast(0.999,tf.float32)
  
         TIB = nr_nuis_gamma * tf.math.pow(self.drift_field, nr_nuis_delta) * pow(self.density / XENON_REF_DENSITY, 0.3)
         Qy = 1. / (TIB * tf.math.pow(energy + nr_nuis_epsilon, nr_nuis_p))
